+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Importing M2 models into Blender using the wow.export add-on now correctly handles transparency and other basic shader settings (requires exported metadata).
 - Added UV animation support to the wow.export Blender add-on (requires exported metadata).
 - Added "Show Bones" option to preview model armature in the model viewer.
+- Maps can now be filtered by expansion using expansion icons above the map list.
 
 ### Data Tables
 - Added "Data" tab that allows you to view data tables within the client!